requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
pytest>=7.4.0
watchdog
